Update aria-label handling in the nav menu
This installs the `@wordpress/scripts` NPM package, adds a `build:js`
script, reogranizes assets in the `/js` directory (code requiring a build
go in the `/src` directory, build files output in `/build`), and finally,
moves the `aria-label` values as settings that can be overridden.

The defaults are ready for translation should we get to that point.